requirements.txt
Django==4.2.5
djangorestframework==3.15.1
flake8==7.0.0
bandit==1.7.7
